refactor: rename `Nazuki.VM.Arch232` to `Nazuki.VM.Binary32`
