+++ src/Nazuki/VM/Binary32/Decimal32/BID.hs
@@ -1,4 +1,4 @@
 -- IEEE 754 decimal32 (binary integer decimal)
 -- See https://en.wikipedia.org/wiki/Decimal32_floating-point_format
 
-module Nazuki.VM.Arch232.Decimal32.BID where
+module Nazuki.VM.Binary32.Decimal32.BID where
