+++ src/Nazuki/VM/Binary32/Decimal32/DPD.hs
@@ -1,4 +1,4 @@
 -- IEEE 754 decimal32 (densely packed decimal)
 -- See https://en.wikipedia.org/wiki/Decimal32_floating-point_format
 
-module Nazuki.VM.Arch232.Decimal32.DPD where
+module Nazuki.VM.Binary32.Decimal32.DPD where
