+++ test/Nazuki/VM/Binary32/Int32Spec.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE BlockArguments #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module Nazuki.VM.Arch232.Int32Spec (spec) where
+module Nazuki.VM.Binary32.Int32Spec (spec) where
 
 import Control.Monad
 import qualified Data.Text as Text
 import Nazuki.Runtime.Runner (run)
-import qualified Nazuki.VM.Arch232.Int32 as Int32
+import qualified Nazuki.VM.Binary32.Int32 as Int32
 import Nazuki.VM.Core (generate)
 import Test.Hspec
